Added tests for avatars. Added in tests plus some fixtures for carrierwave. Moved over configuration for carrierwave test to test_helper.rb.
The tests check that it is possible to
A) delete the avatar file for someone when you delete their user
B) upload an avatar to a user.

Note, when working with fixtures and files, you need to first put the
file plus the complete path in fixtures/files/uploads. The helper will
then copy it to carrierwave test directory, which is a temporary
directory that changes files during the test. This will keep the
fixture's files safe for repeat testing.