Made job applicants unable to be made if job does not accept applicants.
For now, the create job application button exists, it just brings up a notification that you can't apply to the job.

Can change whether job accepts applicants or not in the edit page. May want to move this into just one method and call it through a button or something.